requirements.txt
# ============================================
# GreenCloudRL - Dependencies
# ============================================

# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Reinforcement Learning
gymnasium>=0.29.0

# Simulation
simpy>=4.0.0

# Explainability
shap>=0.43.0
captum>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment Tracking
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0
tabulate>=0.9.0

# Testing
pytest>=7.0.0
